requirements.txt
discord.py==2.3.1
wavelink==2.6.0